ip, display api route refresh
